VSC-ctrl-drag-copy.ahk - Ctrl+LMouse-click drag copy support for Visual Studio Code.
Goal: As of this time (10/18/2016) VSC doesn't support copying selected
text, by
holding down the Ctrl key and dragging the cursor to a 'paste' location.

In a vast number of apps (PowershellISE, WinWord, notepad++, notepad2
etc), this is a standard mouse shortcut.
I don't doubt there's a native VisualStudioCode keybinding etc, or
extension to do this, but I wanted something quickly today.
So I quickly slapped this together. Not a lot of testing, but it seems
to work for me in VSC v1.6.1.